Factor out Lobby component
