+++ src/common/containers/Game.js
@@ -2,19 +2,15 @@ import React, { Component } from 'react';
 import Helmet from 'react-helmet';
 import Paper from 'material-ui/lib/paper';
 import RaisedButton from 'material-ui/lib/raised-button';
-import TextField from 'material-ui/lib/text-field';
-import SelectField from 'material-ui/lib/select-field';
-import MenuItem from 'material-ui/lib/menus/menu-item';
 import { connect } from 'react-redux';
-import { isNil, shuffle } from 'lodash';
+import { isNil } from 'lodash';
 
-import { SHUFFLE_DECKS } from '../constants';
-import { instantiateCard } from '../util/common';
 import { getAttribute } from '../util/game';
 import Board from '../components/game/Board';
+import CardViewer from '../components/game/CardViewer';
+import Lobby from '../components/game/Lobby';
 import PlayerArea from '../components/game/PlayerArea';
 import Status from '../components/game/Status';
-import CardViewer from '../components/game/CardViewer';
 import VictoryScreen from '../components/game/VictoryScreen';
 import * as gameActions from '../actions/game';
 import * as socketActions from '../actions/socket';
@@ -50,9 +46,7 @@ export function mapStateToProps(state) {
     blueDeck: state.game.players.blue.deck,
     orangeDeck: state.game.players.orange.deck,
 
-    numPlayersOnline: state.socket.numPlayersOnline,
-    waitingPlayers: state.socket.waitingPlayers,
-    hosting: state.socket.hosting,
+    socket: state.socket,
     availableDecks: state.collection.decks,
 
     sidebarOpen: state.layout.present.sidebarOpen
@@ -101,15 +95,6 @@ export function mapDispatchToProps(dispatch) {
 }
 
 export class Game extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      gameName: '',
-      selectedDeck: 0
-    };
-  }
-
   isMyTurn() {
     return this.props.currentTurn === this.props.player;
   }
@@ -199,119 +184,49 @@ export class Game extends Component {
     );
   }
 
-  renderLobby() {
-    const paperStyle = {padding: 20, marginBottom: 20, position: 'relative'};
-    const buttonStyle = {position: 'absolute', top: 0, bottom: 0, right: 20, margin: 'auto', color: 'white'};
-
-    const [numPlayersElt, waitingElt, chooseDeckElt, joinGamesElt, startNewGameElt] = [
-      <Paper style={paperStyle}>
-        <div>{this.props.numPlayersOnline} player(s) online</div>
-      </Paper>,
-
-      <Paper style={paperStyle}>
-        <div>Waiting for an opponent ...</div>
-      </Paper>,
-
-      <Paper style={paperStyle}>
-        <SelectField
-          value={this.state['selectedDeck']}
-          floatingLabelText="Choose a deck"
-          style={{width: '80%', marginRight: 25}}
-          onChange={(e, idx, value) => {
-            this.setState(state => state.selectedDeck = idx);
-          }}>
-          {this.props.availableDecks.map((deck, idx) =>
-            <MenuItem key={idx} value={idx} primaryText={`${deck.name} (${deck.cards.length} cards)`}/>
-          )}
-        </SelectField>
-      </Paper>,
-
-      (this.props.waitingPlayers || []).map(game =>
-        <Paper style={paperStyle} key={game.id}>
-          <div>
-            <b>{game.name}</b>
-          </div>
-          <RaisedButton
-            secondary
-            label="Join Game"
-            style={buttonStyle}
-            onTouchTap={e => {
-              const deck = this.props.availableDecks[this.state.selectedDeck].cards.map(instantiateCard);
-              this.props.onJoinGame(game.id, SHUFFLE_DECKS ? shuffle(deck) : deck);
-            }} />
-        </Paper>
-      ),
-
-      <Paper style={paperStyle}>
-        <TextField
-          value={this.state['gameName']}
-          floatingLabelText="Game name"
-          style={{width: '50%'}}
-          onChange={e => { this.setState({gameName: e.target.value}); }} />
-        <RaisedButton
-          secondary
-          disabled={this.state.gameName === ''}
-          label="Host New Game"
-          style={buttonStyle}
-          onTouchTap={e => {
-            const deck = this.props.availableDecks[this.state.selectedDeck].cards.map(instantiateCard);
-            this.props.onHostGame(this.state.gameName, SHUFFLE_DECKS ? shuffle(deck) : deck);
-          }} />
-      </Paper>
-    ];
-
-    if (this.props.hosting) {
+  renderGameArea() {
+    if (this.props.started) {
       return (
-        <div>
-          {numPlayersElt}
-          {waitingElt}
-        </div>
+        <Paper style={{padding: 20, position: 'relative'}}>
+          {this.renderPlayerArea('orange')}
+
+          <div style={{position: 'relative'}}>
+            <CardViewer hoveredCard={this.hoveredCard()} />
+            <Status
+              currentTurn={this.props.currentTurn}
+              status={this.isMyTurn() ? this.props.status : {}} />
+            <Board
+              player={this.props.player}
+              currentTurn={this.props.currentTurn}
+              selectedTile={this.props.selectedTile}
+              target={this.props.target}
+              bluePieces={this.props.bluePieces}
+              orangePieces={this.props.orangePieces}
+              playingCardType={this.props.playingCardType}
+              onSelectTile={(hexId, action, intmedMoveHexId) => this.onSelectTile(hexId, action, intmedMoveHexId)}
+              onHoverTile={(hexId, action) => this.onHoverTile(hexId, action)} />
+            <RaisedButton
+              secondary
+              disabled={!this.isMyTurn()}
+              label="End Turn"
+              style={{position: 'absolute', top: 0, bottom: 0, right: 0, margin: 'auto', color: 'white'}}
+              onTouchTap={this.props.onPassTurn} />
+          </div>
+
+          {this.renderPlayerArea('blue')}
+
+          <VictoryScreen winner={this.props.winner} onClick={this.props.onClick} />
+        </Paper>
       );
     } else {
       return (
-        <div>
-          {numPlayersElt}
-          {chooseDeckElt}
-          {joinGamesElt}
-          {startNewGameElt}
-        </div>
+        <Lobby
+          socket={this.props.socket}
+          availableDecks={this.props.availableDecks}
+          onHostGame={this.props.onHostGame}
+          onJoinGame={this.props.onJoinGame} />
       );
     }
-  }
-
-  renderGameArea() {
-    return (
-      <Paper style={{padding: 20, position: 'relative'}}>
-        {this.renderPlayerArea('orange')}
-
-        <div style={{position: 'relative'}}>
-          <CardViewer hoveredCard={this.hoveredCard()} />
-          <Status
-            currentTurn={this.props.currentTurn}
-            status={this.isMyTurn() ? this.props.status : {}} />
-          <Board
-            player={this.props.player}
-            currentTurn={this.props.currentTurn}
-            selectedTile={this.props.selectedTile}
-            target={this.props.target}
-            bluePieces={this.props.bluePieces}
-            orangePieces={this.props.orangePieces}
-            playingCardType={this.props.playingCardType}
-            onSelectTile={(hexId, action, intmedMoveHexId) => this.onSelectTile(hexId, action, intmedMoveHexId)}
-            onHoverTile={(hexId, action) => this.onHoverTile(hexId, action)} />
-          <RaisedButton
-            secondary
-            disabled={!this.isMyTurn()}
-            label="End Turn"
-            style={{position: 'absolute', top: 0, bottom: 0, right: 0, margin: 'auto', color: 'white'}}
-            onTouchTap={this.props.onPassTurn} />
-        </div>
-
-        {this.renderPlayerArea('blue')}
-
-        <VictoryScreen winner={this.props.winner} onClick={this.props.onClick} />
-      </Paper>
-    );
   }
 
   render() {
@@ -321,7 +236,7 @@ export class Game extends Component {
         margin: '48px 72px'
       }}>
         <Helmet title="Game"/>
-        {this.props.started ? this.renderGameArea() : this.renderLobby()}
+        {this.renderGameArea()}
       </div>
     );
   }
@@ -352,9 +267,7 @@ Game.propTypes = {
   blueDeck: array,
   orangeDeck: array,
 
-  numPlayersOnline: number,
-  waitingPlayers: array,
-  hosting: bool,
+  socket: object,
   availableDecks: array,
 
   selectedCard: number,
